+++ src/model.cpp
@@ -82,7 +82,7 @@ std::vector<uint8_t> Model:: getMatrix(){
 	return this->cell.getMatrix();
 }
 
-std::array<double,3> Model::getResolution(){
+std::array<unsigned int,3> Model::getResolution(){
 	return this->cell.getResolution();
 }
